0-promise.js

+++ 0x01-ES6_promise/.eslintrc.js
@@ -20,7 +20,7 @@ module.exports = {
     "no-restricted-syntax": ["error", "LabeledStatement", "WithStatement"],
     quotes: [
       "error",
-      "double",
+      "single",
       {
         allowTemplateLiterals: true,
         avoidEscape: true,
